feat: faux HTTP proxy TLS support
When `/tls/keystore.jks` exists this signals to the implementation that
the service listens on port 8443 with HTTP/TLS. This will not bring
about support for HTTP/TLS proxy tunneling via `CONNECT` -- it merely
configures the different port and enables the TLS support.

Changes to the OpenShift configuration enable OpenShift specific
issuance of the key/certificate for the service.

Sadly, FMP enricher for TLS works only partially so additional
configuration was required, that would usually be done by FMP to
configure the DeploymentConfig.